+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [1.0.2] - 2023-04-15
+
+### Added
+
+- Validation for prop `customSettings`.
+- Customized log messages `msg`, `warn` & `error`.
+- `reset` button in editor.
+
+### Changed
+
+- Name of `$emit` event `clicked`. Now called `useSettings`.
+- Folder structure in editor.
+
 ## [1.0.1] - 2023-04-13
 
 ### Added